# [#419] Improve performance of Java source language