[#419] Improve performance of Java source language

The patch is based on batched compilation of Java source files. Most of the time for the Java source language is eaten up by repetitive compilation of referenced Java files, if we compile each Java source file individually.
This is changed towards a queue of not-yet-compiled, fully installed Java source resources. As soon as a resource is compiled, it compiles all of its friends and caches the result. A subsequent request to fully install Java source resource contents will first consult the cache and reuse the available cached data, if any.