@kthoms 2019-10-29
The speedup is significant on my box:
```
BEFORE
100 file took 330 ms
200 file took 235 ms
400 file took 339 ms
800 file took 424 ms
1600 file took 540 ms
3200 file took 1102 ms
6400 file took 2109 ms
12800 file took 3638 ms
25600 file took 6600 ms
51200 file took 11674 ms

AFTER
100 file took 207 ms
200 file took 101 ms
400 file took 161 ms
800 file took 215 ms
1600 file took 266 ms
3200 file took 307 ms
6400 file took 511 ms
12800 file took 1225 ms
25600 file took 2809 ms
51200 file took 5009 ms
```